Add option to provide intent action and data uri
When invoking an activity it is possible to provide an Android action
and data uri. Also, when providing the package name and Activity name
it is now possible to provide a fully qualified class name. When the
name of the Activity is prefixed with a '.'; then the Activity name
is automatically prefixed with the package name.